third_party/WebKit/LayoutTests/media/track/track-css-all-cues-expected.txt

Test that style to all cues is applied correctly.
EVENT(canplaythrough)
EVENT(seeked)
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).color == 'rgb(128, 0, 128)') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).backgroundColor == 'rgb(0, 255, 0)') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).textDecorationLine == 'underline') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).textDecorationStyle == 'dashed') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).textDecorationColor == 'rgb(0, 255, 255)') OK
END OF TEST
